src/legacy-langchain/requirements.txt
# langchain-openai
langchain==0.1.16
openai
langchain-openai
python-dotenv